refactor(upload-modal): move components for file, Google Drive, link, and text uploads with improved UI and error handling
